cleancode

+++ src/main/java/LeetCode1.java
@@ -25,14 +25,14 @@ public class LeetCode1 {
             });
         }
         threadPool.shutdown();
-        System.out.println((int)'a');
-        System.out.println((int)'A');
+        System.out.println((int) 'a');
+        System.out.println((int) 'A');
     }
 
     public int maxDifference(String s) {
         int maxj = 0, minj = 100, maxo = 0, mino = 100, cur = 0;
         for (int i = 0; i < 26; i++) {
-            cur = countChar(s, (char) ('a' + i));
+            cur = (int) countChar(s, (char) ('a' + i));
             if (cur % 2 == 1) {
                 maxj = Math.max(maxj, cur);
                 minj = Math.min(minj, cur);
